+++ server/data/users.js
@@ -1,0 +1,8 @@
+const users = [
+  {
+    id: 1,
+    name: 'Bob Marley',
+  },
+];
+
+module.exports = users;